Added EnterZone and Location Event parsing and event firing.

+++ parser/src/main/java/com/pyrese/eq/parser/events/LocationEvent.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class LocationEvent extends LogEvent {
-    int y;
-    int x;
-    int z;
+    double y;
+    double x;
+    double z;
 }
